Finished Creature Creation logic and scene switching
Finished Creature Creation logic, Which includes a new scene that is created to hold the creatures in the intermediary step between the creature creation and putting it into the initiative list.
As such, scene switching needed to change to accommodate for this. Now, the scenes are loaded additively so that scenes don't get removed immediately.
Two more tests were also written, both of which are playmode tests, one checking for the lack of a creature being created when not all inputs are full, and another that checks if a creature is created when all inputs are filled.