Changed field colors on character creation, Initiative control changes
Changed all the field colors on character creation to be more consistent with the rest of the scenes (when hovered over they also appear a different color).

In the PreviousTurn Method  added the ability to return to the final turn in the previous round of combat if trying to click the button while on the first initiative in the list. (Will not work if it's the first round of combat).

Lastly, added a new method to ControlCombat called SortInitiative. Currently implements a bubble search to sort the list in the hierarchy in descending order based on initiative (does not change their place on the screen, gotta work on that still).